Fixed minor highlight issues.
Only highlight the first segment of a line (if necessary) and never highlight its leading indentation. We don't want every ANSI color change to be a separately highlighted segment in a line.

Also, only use the unique output Guid (if any) at the beginning of a line and not on its indentation or ANSI color segments.